+++ docs/CHANGELOG.rst
@@ -58,6 +58,7 @@ Fixed
 * Fix ``htseq_count`` and ``featureCounts`` for large files
 * Fix ``upload gtf annotation``
 * Fix gene_id field type for differential expression storage objects
+* Order data objects in ``SampleViewSet``
 
 
 ================